ReceiptParser/requirements.txt
pytesseract
Pillow
pdf2image
SQLAlchemy
click
python-dotenv
ollama
customtkinter